+++ LiteXLoader/LiteXLoader/API/EntityAPI.h
@@ -36,6 +36,7 @@ public:
 
     Local<Value> teleport(const Arguments& args);
     Local<Value> kill(const Arguments& args);
+    Local<Value> isPlayer(const Arguments& args);
     Local<Value> toPlayer(const Arguments& args);
     Local<Value> setOnFire(const Arguments& args);